reorganizacja kodu, tworzenie/odtwarzanie bazy danych, odczytywanie jednej wartości z bazy
