--- _config.yml
@@ -1,1 +1,0 @@
-theme: jekyll-theme-time-machine
