+++ requirements.txt
@@ -1,0 +1,8 @@
+Click==7.0
+Flask==1.0.2
+future==0.17.1
+itsdangerous==1.1.0
+Jinja2==2.10
+MarkupSafe==1.1.0
+pyglet==1.3.2
+Werkzeug==0.14.1
