+++ zeton/data_access.py
@@ -1,0 +1,39 @@
+import json
+from datetime import datetime, date
+
+
+def json_serial(obj):
+    """JSON serializer for objects not serializable by default json code"""
+
+    if isinstance(obj, (datetime, date)):
+        return obj.isoformat()
+    raise TypeError("Type %s not serializable" % type(obj))
+
+
+def wczytaj_dane():
+    """
+    Wczytuje dane z pliku dane.json
+    Jeśli nie ma tego pliku to go tworzy z ustawionymi wartościmai na "0"
+    :return: dane ucznia
+    """
+    try:
+        with open('dane.json', 'r') as plik:
+            dane = json.load(plik)
+    except FileNotFoundError:
+        with open('dane.json', 'w') as plik:
+            dane = {"punkty": 0, "szkolny_rekord_tygodnia": 0, "ban": False}
+            json.dump(dane, plik, default=json_serial)
+    return dane
+
+
+def zapisz_dane(dane):
+    """
+    zapisuje dane do pliku "dane.json"
+    :param dane: dane ucznia - słownik (dictionary)
+    :return: None
+    """
+    try:
+        with open('dane.json', 'w') as plik:
+            json.dump(dane, plik, default=json_serial)
+    except:
+        return f'Nie można zapisać dancyh do pliku'
